1.[DELETE]:delete Byteorderconverter 2.[MOVE]:move byteorder code to msgnode
