+++ chatting-server/include/buffer/MsgNode.hpp
@@ -2,7 +2,7 @@
 #pragma once
 #ifndef _MSGNODE_H_
 #define _MSGNODE_H_
-#include <buffer/ByteOrderConverter.hpp>
+#include <boost/asio/detail/socket_ops.hpp>
 #include <cstdint>
 #include <functional>
 #include <iterator>
@@ -69,6 +69,46 @@ struct recv_msg_check<
                 decltype(std::declval<std::decay_t<Container>>().begin()),
                 decltype(std::declval<std::decay_t<Container>>().end())>>
     : public std::true_type {};
+
+
+#include <boost/asio/detail/socket_ops.hpp>
+#include <type_traits>
+
+template <typename T,
+          typename std::enable_if_t<sizeof(T) == sizeof(uint16_t), int> = 0>
+T convert_from_network(T value) {
+          return boost::asio::detail::socket_ops::network_to_host_short(value);
+}
+
+template <typename T,
+          typename std::enable_if_t<sizeof(T) == sizeof(uint32_t), int> = 0>
+T convert_from_network(T value) {
+          return boost::asio::detail::socket_ops::network_to_host_long(value);
+}
+
+template <typename T,
+          typename std::enable_if_t<sizeof(T) == sizeof(uint16_t), int> = 0>
+T convert_to_network(T value) {
+          return boost::asio::detail::socket_ops::host_to_network_short(value);
+}
+
+template <typename T,
+          typename std::enable_if_t<sizeof(T) == sizeof(uint32_t), int> = 0>
+T convert_to_network(T value) {
+          return boost::asio::detail::socket_ops::host_to_network_long(value);
+}
+
+struct ByteOrderConverter {
+          template <typename T> T operator()(T value) const {
+                    return convert_from_network(value);
+          }
+};
+
+struct ByteOrderConverterReverse {
+          template <typename T> T operator()(T value) const {
+                    return convert_to_network(value);
+          }
+};
 
 template <typename Container> struct MsgHeader {
   /*letting tcpnetwork to handle protected _buffer*/
